backend/requirements.txt
Flask==2.0.1
psycopg2-binary
covalent
qiskit==0.40.0
qiskit-ibm-runtime==0.8.0
pandas
scikit-learn
sqlalchemy
